refactor: 카테고리의 상세내역 화면을 구현했습니다.

+++ common/src/main/java/com/paykidscompose/common/usecase/allowance/expense/DeleteExpenseUseCase.kt
@@ -18,6 +18,6 @@ class DeleteExpenseUseCase(
     }
 
     data class Params(
-        val id: Int
+        val id: Long
     )
 }
